Rework command initialization
This patch provides new infrastructure for the commands of px so
they can be used in normal operation and in the unit tests. The
main issue with the normal model is that the unit tests cannot
re-initialize the variables since they are initialized as global
variables and with the `init()` functions.

This patch provides a simple method for initializing variables
and the command flags by providing two functions:

* RegisterCommandVar
* RegisterCommandInit

These functions can be use the same model as used by Ginkgo. Here
is an example how they can be used:

```
var createCmd *cobra.Command

var _ = RegisterCommandVar(func() {
	createCmd = &cobra.Command{
		Use:   "create",
		Short: "Create an object in Portworx",
		Run: func(cmd *cobra.Command, args []string) {
			util.Printf("Please see px create "+
			"--help for more information")
		},
	}
})

var _ = RegisterCommandInit(func() {
	rootCmd.AddCommand(createCmd)
})
```

This patch also adds `cmd/main.go` to provide the appropriate
functions to re-initialize the command variables.